# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloudimart api Documentation",
        "description": "Cloudimart api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "local server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully."
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Product name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "image": {
                                        "description": "Product image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_id": {
                                        "description": "Product category",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product Added"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/categories/{categoryId}/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products by category",
                "operationId": "4b267b0def568f354a189242b9a19b0a",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Id of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products retrieved."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}